Consolidate `utils.rs` in babylon contract
