+++ contracts/babylon/src/ibc.rs
@@ -222,11 +222,7 @@ pub(crate) mod ibc_packet {
         btc_headers: &BtcHeaders,
     ) -> StdResult<IbcReceiveResponse> {
         // Submit headers to BTC light client
-        let msg = crate::utils::btc_light_client_executor::new_btc_headers_msg(
-            deps,
-            &btc_headers.headers,
-        )
-        .map_err(|e| {
+        let msg = crate::utils::new_btc_headers_msg(deps, &btc_headers.headers).map_err(|e| {
             let err = format!("CONTRACT: handle_btc_headers, failed to submit BTC headers: {e}");
             deps.api.debug(&err);
             StdError::generic_err(err)
