--- contracts/babylon/src/utils/mod.rs
@@ -1,2 +1,0 @@
-pub mod btc_light_client_executor;
-pub mod btc_light_client_querier;
